:rocket: follow request

+++ src/main/java/com/esgi/api_project_annuel/web/request/FollowRequest.java
@@ -1,7 +1,6 @@
 package com.esgi.api_project_annuel.web.request;
 
 public class FollowRequest {
-    public long followerUserId;
-
-    public long followedUserId;
+    public int followerUserId;
+    public int followedUserId;
 }
